#Add repo location commandline option added
#Add error catching in case missing config files and git folder
#Add add appveyor ci building

+++ src/ChangelogGenerator.Core/ChangelogCore.cs
@@ -17,8 +17,18 @@ namespace ChangelogGenerator.Core
                 .MapResult(o => GenerateChangelog(o), _ => 1);
         }
 
-        public void GenerateChangelog(ChangelogSettings settings)
+        public int GenerateChangelog(ChangelogSettings settings)
         {
+            var gitFolder = Path.Combine(settings.GitRepoLocation, ".git");
+
+            if (!Directory.Exists(gitFolder))
+            {
+                Console.Error.WriteLine($"Could not detect .git folder in {settings.GitRepoLocation}");
+                Console.Error.WriteLine($"Without a git folder the history can not be analyzed");
+                return 1;
+            }
+
+
             GitChangelog changelog = GitChangelog.Generate(settings);
 
             var content = changelog.GetChangeLogText();
@@ -31,6 +41,8 @@ namespace ChangelogGenerator.Core
             {
                 File.WriteAllText(settings.ChangelogLocation, content);
             }
+
+            return 0;
         }
 
         private int GenerateChangelog(ChangelogCommandlineSettings commandLineSettings)
@@ -41,11 +53,27 @@ namespace ChangelogGenerator.Core
             {
                 ChangelogSettings newSettings = new ChangelogSettings();
 
+                var directory = Path.GetDirectoryName(configLocation);
+
+                if (!Directory.Exists(directory))
+                {
+                    Console.Error.WriteLine($"Path {directory} does not exist");
+                    return 1;
+                }
+                    
+
                 var content = JsonConvert.SerializeObject(newSettings, Formatting.Indented);
 
                 File.WriteAllText(configLocation, content);
 
                 return 0;
+            }
+
+            if(!File.Exists(configLocation))
+            {
+                Console.Error.WriteLine($"Config file {configLocation} does not exist");
+                Console.Error.WriteLine("Create a config file using -m or --make-config");
+                return 1;
             }
 
             ChangelogSettings settings = JsonConvert.DeserializeObject<ChangelogSettings>(File.ReadAllText(configLocation));
@@ -60,9 +88,12 @@ namespace ChangelogGenerator.Core
                 settings.UnreleasedTitle = commandLineSettings.NextVersion;
             }
 
-            GenerateChangelog(settings);
+            if (!string.IsNullOrEmpty(commandLineSettings.RepoLocation))
+            {
+                settings.GitRepoLocation = commandLineSettings.RepoLocation;
+            }
 
-            return 0;
+            return GenerateChangelog(settings);
         }
     }
 }
